Sensors (#7)
* Refactor moisture and battery sensor handling for improved accuracy and type consistency

* Refactor moisture sensor data handling for improved type consistency and average calculation

* Remove unused dependencies and update README for clarity

* Replace defmt with log for logging functionality and update dependencies in Cargo.toml

+++ src/domain.rs
@@ -6,7 +6,7 @@ use alloc::{
 use core::fmt::{Display, Formatter, Result};
 use serde::{Deserialize, Serialize};
 
-const WATER_LEVEL_THRESHOLD: u32 = 3000;
+const WATER_LEVEL_THRESHOLD: u16 = 3000;
 
 /// Struct to hold sensor data
 #[derive(Default, Debug)]
@@ -50,8 +50,8 @@ pub enum Sensor {
     PumpTrigger(bool),
 }
 
-impl From<u32> for WaterLevel {
-    fn from(value: u32) -> Self {
+impl From<u16> for WaterLevel {
+    fn from(value: u16) -> Self {
         if value < WATER_LEVEL_THRESHOLD {
             WaterLevel::Empty
         } else {
